Implement proxy and XMP handling for peer-to-peer synchronization
- Added proxy subscription and download functionality in `proxy.go` to manage proxy announcements and fetch proxies from peers.
- Implemented XMP handling in `xmp.go`, including inbound XMP processing, pushing updates to peers, and serving XMP files over HTTP.
- Introduced methods for merging XMP metadata to preserve existing edits while applying new updates.
- Enhanced logging for better traceability of proxy and XMP operations.
