+++ src/p2p/daemon.go
@@ -1,25 +1,18 @@
 package main
 
 import (
-	"bytes"
 	"context"
-	"crypto/ed25519"
-	"crypto/rand"
 	"crypto/sha256"
 	"crypto/tls"
 	"crypto/x509"
-	"crypto/x509/pkix"
-	"database/sql"
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
 	"io"
 	"log"
-	"math/big"
 	"net"
 	"net/http"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"regexp"
 	"sort"
@@ -28,7 +21,6 @@ import (
 	"time"
 
 	gocrypto "crypto"
-	"net/url"
 
 	libp2p "github.com/libp2p/go-libp2p"
 	pubsub "github.com/libp2p/go-libp2p-pubsub"
@@ -36,12 +28,8 @@ import (
 	"github.com/libp2p/go-libp2p/core/host"
 	"github.com/libp2p/go-libp2p/core/peer"
 	"github.com/libp2p/go-libp2p/p2p/discovery/mdns"
-	multiaddr "github.com/multiformats/go-multiaddr"
 	"golang.org/x/crypto/pbkdf2"
 
-	"github.com/huin/goupnp/dcps/internetgateway1"
-	"github.com/huin/goupnp/dcps/internetgateway2"
-	natpmp "github.com/jackpal/go-nat-pmp"
 	_ "modernc.org/sqlite"
 )
 
@@ -55,212 +43,6 @@ const (
 	// passphraseHeader is required on every inbound HTTP request.
 	passphraseHeader = "X-DT-Auth"
 )
-
-// ---------------------------------------------------------------------------
-// TLS helpers
-// ---------------------------------------------------------------------------
-
-// tlsCertFromKey derives a self-signed TLS certificate from the p2p identity
-// key (itself derived from the shared passphrase).  All peers with the same
-// passphrase share the same key and therefore the same certificate fingerprint,
-// so the fingerprint doubles as a shared secret for peer identification.
-func tlsCertFromKey(priv crypto.PrivKey) (tls.Certificate, error) {
-	raw, err := priv.Raw()
-	if err != nil {
-		return tls.Certificate{}, fmt.Errorf("key raw: %w", err)
-	}
-	var edPriv ed25519.PrivateKey
-	switch len(raw) {
-	case ed25519.PrivateKeySize: // 64 bytes — full key
-		edPriv = ed25519.PrivateKey(raw)
-	case ed25519.SeedSize: // 32 bytes — seed only
-		edPriv = ed25519.NewKeyFromSeed(raw)
-	default:
-		return tls.Certificate{}, fmt.Errorf("unexpected Ed25519 key length %d", len(raw))
-	}
-
-	serial, err := rand.Int(rand.Reader, new(big.Int).Lsh(big.NewInt(1), 128))
-	if err != nil {
-		return tls.Certificate{}, err
-	}
-	tmpl := &x509.Certificate{
-		SerialNumber: serial,
-		Subject:      pkix.Name{CommonName: "darktable-p2p"},
-		NotBefore:    time.Now().Add(-time.Hour),
-		NotAfter:     time.Now().Add(100 * 365 * 24 * time.Hour),
-		KeyUsage:     x509.KeyUsageDigitalSignature,
-		ExtKeyUsage:  []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth, x509.ExtKeyUsageClientAuth},
-	}
-	certDER, err := x509.CreateCertificate(rand.Reader, tmpl, tmpl, edPriv.Public(), edPriv)
-	if err != nil {
-		return tls.Certificate{}, fmt.Errorf("create cert: %w", err)
-	}
-	return tls.Certificate{
-		Certificate: [][]byte{certDER},
-		PrivateKey:  edPriv,
-	}, nil
-}
-
-// pubkeyFingerprint returns the lowercase hex SHA256 of the certificate's
-// SubjectPublicKeyInfo (SPKI) DER bytes — stable across certificate re-issuance.
-func pubkeyFingerprint(cert *x509.Certificate) string {
-	spki, err := x509.MarshalPKIXPublicKey(cert.PublicKey)
-	if err != nil {
-		// Fallback: hash the raw SPKI field embedded in the cert.
-		h := sha256.Sum256(cert.RawSubjectPublicKeyInfo)
-		return hex.EncodeToString(h[:])
-	}
-	h := sha256.Sum256(spki)
-	return hex.EncodeToString(h[:])
-}
-
-// passphraseToken derives a fixed-length authentication token from the
-// passphrase.  Sent as an HTTP header; TLS encryption protects it in transit.
-func passphraseToken(passphrase string) string {
-	h := sha256.Sum256([]byte("darktable-p2p-auth\x00" + passphrase))
-	return hex.EncodeToString(h[:])
-}
-
-// peerKeysPath returns the path of the allowed-fingerprints file.
-func peerKeysPath() string {
-	if dir := os.Getenv("XDG_CONFIG_HOME"); dir != "" {
-		return dir + "/darktable/peer.keys"
-	}
-	if home, err := os.UserHomeDir(); err == nil {
-		return home + "/.config/darktable/peer.keys"
-	}
-	return ""
-}
-
-// loadAllowedKeyHashes returns the set of accepted server TLS public key
-// fingerprints.  ownFP is always included so nodes on the same passphrase
-// accept each other by default.  Additional fingerprints may be listed in
-// ~/.config/darktable/peer.keys, one per line (plain hex or "sha256:<hex>").
-func loadAllowedKeyHashes(ownFP string) map[string]bool {
-	allowed := map[string]bool{ownFP: true}
-	data, err := os.ReadFile(peerKeysPath())
-	if err != nil {
-		return allowed
-	}
-	for _, line := range strings.Split(string(data), "\n") {
-		line = strings.ToLower(strings.TrimSpace(line))
-		if line == "" || strings.HasPrefix(line, "#") {
-			continue
-		}
-		line = strings.TrimPrefix(line, "sha256:")
-		if len(line) == 64 { // SHA256 = 32 bytes = 64 hex chars
-			allowed[line] = true
-		}
-	}
-	return allowed
-}
-
-// newServerTLSConfig returns a TLS config for the HTTP server.
-// Client certificates are not required; authentication is via the
-// passphraseHeader checked at the HTTP handler layer.
-func newServerTLSConfig(cert tls.Certificate) *tls.Config {
-	return &tls.Config{
-		Certificates: []tls.Certificate{cert},
-		MinVersion:   tls.VersionTLS13,
-		ClientAuth:   tls.NoClientCert,
-	}
-}
-
-// newClientTLSConfig returns a TLS config for outbound peer connections.
-// Chain validation is skipped; instead the server's public key fingerprint is
-// checked against the daemon's live allowedKeyFPs map under its mutex, so newly
-// accepted peers are trusted immediately without restarting the daemon.
-func (d *daemon) newClientTLS() *tls.Config {
-	return &tls.Config{
-		Certificates:       []tls.Certificate{d.tlsCert},
-		InsecureSkipVerify: true, //nolint:gosec — deliberate; fingerprint-pinned below
-		VerifyPeerCertificate: func(rawCerts [][]byte, _ [][]*x509.Certificate) error {
-			for _, raw := range rawCerts {
-				c, err := x509.ParseCertificate(raw)
-				if err != nil {
-					continue
-				}
-				fp := pubkeyFingerprint(c)
-				d.allowedKeyFPsMu.RLock()
-				ok := d.allowedKeyFPs[fp]
-				d.allowedKeyFPsMu.RUnlock()
-				if ok {
-					return nil
-				}
-			}
-			return fmt.Errorf("server TLS public key not in allowed list")
-		},
-	}
-}
-
-// probePeerFingerprint dials the peer's HTTPS endpoint without certificate
-// verification and returns the server's public key fingerprint.  Used to
-// identify untrusted peers so the user can choose to accept them.
-func probePeerFingerprint(rawURL string) string {
-	u, err := url.Parse(rawURL)
-	if err != nil || u.Scheme != "https" {
-		return ""
-	}
-	host := u.Host
-	if _, _, err := net.SplitHostPort(host); err != nil {
-		host = fmt.Sprintf("%s:%d", host, proxyHTTPPort)
-	}
-	conn, err := tls.DialWithDialer(
-		&net.Dialer{Timeout: 3 * time.Second}, "tcp", host,
-		&tls.Config{InsecureSkipVerify: true}, //nolint:gosec — fingerprint probe only
-	)
-	if err != nil {
-		return ""
-	}
-	defer conn.Close()
-	certs := conn.ConnectionState().PeerCertificates
-	if len(certs) == 0 {
-		return ""
-	}
-	return pubkeyFingerprint(certs[0])
-}
-
-// recordCandidatePeer stores a discovered-but-untrusted peer and persists the
-// list to ~/.config/darktable/peer.candidates so the preferences UI can read it.
-func (d *daemon) recordCandidatePeer(peerURL, fp string) {
-	d.allowedKeyFPsMu.RLock()
-	already := d.allowedKeyFPs[fp]
-	d.allowedKeyFPsMu.RUnlock()
-	if already {
-		return
-	}
-
-	d.candidatesMu.Lock()
-	if d.candidatePeers == nil {
-		d.candidatePeers = make(map[string]candidatePeer)
-	}
-	if _, exists := d.candidatePeers[fp]; !exists {
-		d.candidatePeers[fp] = candidatePeer{URL: peerURL, Fingerprint: fp, SeenAt: time.Now()}
-		log.Printf("[peer] candidate (untrusted) peer %s  fp=%s", peerURL, fp)
-	}
-	snapshot := make([]candidatePeer, 0, len(d.candidatePeers))
-	for _, c := range d.candidatePeers {
-		snapshot = append(snapshot, c)
-	}
-	d.candidatesMu.Unlock()
-
-	if cfgDir, err := os.UserConfigDir(); err == nil {
-		data, _ := json.Marshal(snapshot)
-		_ = os.WriteFile(filepath.Join(cfgDir, "darktable", "peer.candidates"), data, 0644)
-	}
-}
-
-// requireAuth is HTTP middleware that rejects requests without the correct
-// passphrase token in passphraseHeader.
-func requireAuth(next http.Handler, token string) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.Header.Get(passphraseHeader) != token {
-			http.Error(w, "unauthorized", http.StatusUnauthorized)
-			return
-		}
-		next.ServeHTTP(w, r)
-	})
-}
 
 // reDatetime matches exif:DateTimeOriginal (and similar) in XMP XML.
 var reDatetime = regexp.MustCompile(
@@ -423,28 +205,6 @@ type manifestResp struct {
 	Paths       []string `json:"paths"`
 	Hash        string   `json:"hash,omitempty"`        // sha256 of sorted path list; clients skip sync when unchanged
 	KnownPeers  []string `json:"known_peers,omitempty"` // HTTP URLs of other known peers
-}
-
-// ---------------------------------------------------------------------------
-// XMP helpers
-// ---------------------------------------------------------------------------
-
-// xmpCaptureDate extracts the first date-time field found in XMP XML content.
-func xmpCaptureDate(content string) string {
-	m := reDatetime.FindStringSubmatch(content)
-	if len(m) > 1 {
-		return strings.TrimSpace(m[1])
-	}
-	return ""
-}
-
-// xmpDateFromFile reads a .xmp sidecar and returns its capture date, or "".
-func xmpDateFromFile(xmpPath string) string {
-	data, err := os.ReadFile(xmpPath)
-	if err != nil {
-		return ""
-	}
-	return xmpCaptureDate(string(data))
 }
 
 // ---------------------------------------------------------------------------
@@ -669,108 +429,6 @@ func newDaemon(ctx context.Context, socketPath, passphrase, proxyDir, importDir 
 	d.buildLocalIndex()
 
 	return d, nil
-}
-
-// ---------------------------------------------------------------------------
-// External access via UPnP / NAT-PMP
-// ---------------------------------------------------------------------------
-
-// upnpPortMapper is the subset of goupnp WANIPConnection we need.
-type upnpPortMapper interface {
-	GetExternalIPAddress() (string, error)
-	AddPortMapping(string, uint16, string, uint16, string, bool, string, uint32) error
-}
-
-// initExternalAccess tries UPnP then NAT-PMP to discover the public IP and
-// add a port mapping for proxyHTTPPort.  Sets d.externalURL on success.
-func (d *daemon) initExternalAccess() {
-	time.Sleep(2 * time.Second) // wait for network to settle
-
-	var mapper upnpPortMapper
-
-	if clients, _, err := internetgateway2.NewWANIPConnection2Clients(); err == nil && len(clients) > 0 {
-		mapper = clients[0]
-		log.Printf("[upnp] found WANIPConnection2 gateway")
-	} else if clients, _, err := internetgateway2.NewWANIPConnection1Clients(); err == nil && len(clients) > 0 {
-		mapper = clients[0]
-		log.Printf("[upnp] found WANIPConnection1 (ig2) gateway")
-	} else if clients, _, err := internetgateway1.NewWANIPConnection1Clients(); err == nil && len(clients) > 0 {
-		mapper = clients[0]
-		log.Printf("[upnp] found WANIPConnection1 (ig1) gateway")
-	}
-
-	if mapper != nil {
-		extIP, err := mapper.GetExternalIPAddress()
-		if err == nil && extIP != "" && extIP != "0.0.0.0" {
-			_ = mapper.AddPortMapping("", uint16(proxyHTTPPort), "TCP",
-				uint16(proxyHTTPPort), d.localIP(), true, "darktable-p2p", 3600)
-			d.externalURL = fmt.Sprintf("https://%s:%d", extIP, proxyHTTPPort)
-			log.Printf("[upnp] external URL: %s", d.externalURL)
-			go d.renewUPnP(mapper)
-			return
-		}
-	}
-
-	// Fall back to NAT-PMP.
-	gw := d.defaultGateway()
-	if gw == nil {
-		return
-	}
-	pmp := natpmp.NewClient(gw)
-	if ext, err := pmp.GetExternalAddress(); err == nil {
-		extIP := net.IP(ext.ExternalIPAddress[:]).String()
-		if extIP != "" && extIP != "0.0.0.0" {
-			_, _ = pmp.AddPortMapping("tcp", proxyHTTPPort, proxyHTTPPort, 3600)
-			d.externalURL = fmt.Sprintf("https://%s:%d", extIP, proxyHTTPPort)
-			log.Printf("[nat-pmp] external URL: %s", d.externalURL)
-		}
-	}
-}
-
-// renewUPnP refreshes the port mapping every 50 minutes (lease is 3600s).
-func (d *daemon) renewUPnP(mapper upnpPortMapper) {
-	tick := time.NewTicker(50 * time.Minute)
-	defer tick.Stop()
-	for {
-		select {
-		case <-d.ctx.Done():
-			return
-		case <-tick.C:
-			_ = mapper.AddPortMapping("", uint16(proxyHTTPPort), "TCP",
-				uint16(proxyHTTPPort), d.localIP(), true, "darktable-p2p", 3600)
-		}
-	}
-}
-
-// defaultGateway returns the default IPv4 gateway by connecting a UDP socket
-// (no data is sent) and taking the interface's first-octet-1 address.
-func (d *daemon) defaultGateway() net.IP {
-	conn, err := net.Dial("udp4", "8.8.8.8:53")
-	if err != nil {
-		return nil
-	}
-	defer conn.Close()
-	ip := conn.LocalAddr().(*net.UDPAddr).IP.To4()
-	if ip == nil {
-		return nil
-	}
-	gw := make(net.IP, 4)
-	copy(gw, ip)
-	gw[3] = 1
-	return gw
-}
-
-// HandlePeerFound implements mdns.Notifee.
-func (d *daemon) HandlePeerFound(pi peer.AddrInfo) {
-	if pi.ID == d.host.ID() {
-		return
-	}
-	log.Printf("[mdns] discovered peer %s", pi.ID)
-	if err := d.host.Connect(d.ctx, pi); err != nil {
-		log.Printf("[mdns] connect to %s: %v", pi.ID, err)
-	}
-	// HTTP URL is unknown until the peer's proxy announce arrives on gossipsub;
-	// subscribeProxy will touch the DB at that point.
 }
 
 // ---------------------------------------------------------------------------
@@ -1164,239 +822,6 @@ func (d *daemon) handleConn(conn net.Conn) {
 }
 
 // ---------------------------------------------------------------------------
-// XMP subscription
-// ---------------------------------------------------------------------------
-
-// applyInboundXMP processes an inbound XMP update received from gossipsub or a
-// direct HTTP push.  from is a short identifier used only for logging.
-func (d *daemon) applyInboundXMP(x xmpMsg, from string) {
-	d.xmpMu.Lock()
-	last := d.xmpSeen[x.Path]
-	t := time.UnixMilli(x.Mtime)
-	shouldProcess := t.After(last)
-	if !shouldProcess {
-		if _, statErr := os.Stat(x.Path + ".xmp"); os.IsNotExist(statErr) {
-			shouldProcess = true
-		}
-	}
-	if !shouldProcess {
-		d.xmpMu.Unlock()
-		return
-	}
-	d.xmpSeen[x.Path] = t
-	d.xmpMu.Unlock()
-
-	localPath := d.resolveLocalPath(x.Path, x.Filename, x.CaptureDate)
-	if localPath == "" {
-		log.Printf("[xmp] recv: '%s' (capture=%s) from %s — no local file; fetching proxy",
-			x.Filename, x.CaptureDate, from)
-		go d.fetchAndImport(x.Path, x.Content, x.Filename, x.CaptureDate)
-		return
-	}
-	if localPath != x.Path {
-		log.Printf("[xmp] recv: '%s' from %s — path differs, applying to local '%s'",
-			x.Filename, from, localPath)
-	} else {
-		log.Printf("[xmp] recv: '%s' from %s — applying to '%s'",
-			x.Filename, from, localPath)
-	}
-	// If the incoming XMP has no darktable history (thin mobile push) but the
-	// local file does, merge only the rating/colorlabels so the processing
-	// pipeline is not overwritten.
-	contentToWrite := x.Content
-	if !strings.Contains(x.Content, "darktable:history") {
-		if existing, err := os.ReadFile(localPath + ".xmp"); err == nil &&
-			strings.Contains(string(existing), "darktable:history") {
-			contentToWrite = mergeXMPMeta(string(existing), x.Content)
-		}
-	}
-	if err := writeXMP(localPath, contentToWrite); err != nil {
-		log.Printf("[xmp] write '%s': %v", localPath, err)
-	} else {
-		// Invalidate preview caches so the next fetch regenerates them from
-		// the updated XMP rather than serving the pre-edit stale files.
-		invalidatePreviewCache(localPath)
-		result, _ := json.Marshal(map[string]string{"path": localPath})
-		d.broadcast(socketMsg{Type: "xmp_updated", Data: result})
-		// Record suppress entry so any immediate echo from our darktable
-		// doesn't bounce the edit back to the peer who sent it.
-		if x.SenderURL != "" {
-			d.xmpMu.Lock()
-			d.xmpSuppressFrom[localPath] = xmpSuppressEntry{
-				url:   x.SenderURL,
-				until: time.Now().Add(15 * time.Second),
-			}
-			d.xmpMu.Unlock()
-		}
-	}
-}
-
-func (d *daemon) subscribeXMP() {
-	for {
-		m, err := d.xmpSub.Next(d.ctx)
-		if err != nil {
-			return
-		}
-		// m.ReceivedFrom is the last-hop peer, not the originator.
-		// Also check SenderID to drop our own messages that routed back via gossip.
-		if m.ReceivedFrom == d.host.ID() {
-			continue
-		}
-		var x xmpMsg
-		if err := json.Unmarshal(m.Data, &x); err != nil {
-			continue
-		}
-		if x.SenderID == d.host.ID().String() {
-			continue // own message routed back via intermediate peer
-		}
-		d.applyInboundXMP(x, m.ReceivedFrom.ShortString())
-	}
-}
-
-// pushXMPToPeers sends x directly via HTTP POST to every known reachable peer.
-// This ensures delivery even when gossipsub has not yet established peer
-// connections (peers=0).  excludeURL, if non-empty, is skipped — used to avoid
-// bouncing an edit back to the peer we just received it from.
-func (d *daemon) pushXMPToPeers(x xmpMsg, excludeURL string) {
-	myURL := d.localProxyURL()
-	myLANURL := fmt.Sprintf("https://%s:%d", d.localIP(), proxyHTTPPort)
-	var targets []string
-	for _, u := range d.allPeerURLs() {
-		if u != "" && u != myURL && u != myLANURL && u != excludeURL {
-			targets = append(targets, u)
-		}
-	}
-	log.Printf("[xmp] push targets: myURL=%s  targets=%v", myURL, targets)
-	if len(targets) == 0 {
-		return
-	}
-	b, err := json.Marshal(x)
-	if err != nil {
-		return
-	}
-	for _, baseURL := range targets {
-		baseURL := baseURL
-		go func() {
-			resp, err := d.httpPost(baseURL+"/xmp", "application/json", bytes.NewReader(b))
-			if err != nil {
-				log.Printf("[xmp] direct push to %s: %v", baseURL, err)
-				return
-			}
-			resp.Body.Close()
-			if resp.StatusCode == http.StatusNoContent || resp.StatusCode == http.StatusOK {
-				log.Printf("[xmp] direct push to %s: ok", baseURL)
-			} else {
-				log.Printf("[xmp] direct push to %s: HTTP %d", baseURL, resp.StatusCode)
-			}
-		}()
-	}
-}
-
-// fetchAndImport downloads a proxy for remotePath and imports it to darktable.
-// Also writes any XMP content received alongside the proxy announcement.
-func (d *daemon) fetchAndImport(remotePath, xmpContent, filename, captureDate string) {
-	// Determine local destination.
-	localPath := d.localDestination(remotePath)
-
-	// Download proxy: prefer libp2p-known peers, fall back to peersToSync().
-	d.peersMu.RLock()
-	var baseURL string
-	for _, u := range d.peers {
-		baseURL = u
-		break
-	}
-	d.peersMu.RUnlock()
-
-	if baseURL == "" {
-		myURL := d.localProxyURL()
-		myLANURL := fmt.Sprintf("https://%s:%d", d.localIP(), proxyHTTPPort)
-		for _, u := range d.peersToSync() {
-			if u != myURL && u != myLANURL {
-				baseURL = u
-				break
-			}
-		}
-	}
-
-	if baseURL == "" {
-		log.Printf("[import] no peer available to fetch proxy for '%s'", filename)
-		return
-	}
-
-	if !d.downloadProxy(remotePath, localPath, baseURL) {
-		return
-	}
-
-	// Write XMP if provided.
-	if xmpContent != "" {
-		if err := writeXMP(localPath, xmpContent); err != nil {
-			log.Printf("[import] write XMP for '%s': %v", filename, err)
-		}
-	}
-
-	log.Printf("[import] imported '%s' → '%s'", filename, localPath)
-	d.addToLocalIndex(localPath)
-
-	// Broadcast the canonical raw path so darktable imports it as the raw file
-	// and uses 4F9A9030.CR3.xmp for develop settings.  _image_import_internal
-	// will transparently fall back to the .proxy.avif when the raw is absent.
-	result, _ := json.Marshal(map[string]string{"path": localPath})
-	d.broadcast(socketMsg{Type: "image_imported", Data: result})
-}
-
-// ---------------------------------------------------------------------------
-// Proxy subscription
-// ---------------------------------------------------------------------------
-
-func (d *daemon) subscribeProxy() {
-	for {
-		m, err := d.proxySub.Next(d.ctx)
-		if err != nil {
-			return
-		}
-		if m.ReceivedFrom == d.host.ID() {
-			continue
-		}
-		var ann proxyAnnounce
-		if err := json.Unmarshal(m.Data, &ann); err != nil {
-			continue
-		}
-		d.peersMu.Lock()
-		d.peers[m.ReceivedFrom] = ann.BaseURL
-		d.peersMu.Unlock()
-		d.pdb.touch(ann.BaseURL, ann.SenderID) // persist gossipsub-discovered peer
-
-		// If we haven't successfully synced with this peer yet, verify it
-		// immediately so it enters d.syncedPeers and becomes a fetch target.
-		// (Without this, a new peer would wait up to 60 s for the sync ticker.)
-		d.syncedPeersMu.Lock()
-		alreadySynced := d.syncedPeers[ann.BaseURL]
-		d.syncedPeersMu.Unlock()
-		if !alreadySynced {
-			go d.syncWithPeer(ann.BaseURL)
-		}
-
-		log.Printf("[proxy] peer %s announced %d proxies at %s",
-			m.ReceivedFrom.ShortString(), len(ann.Paths), ann.BaseURL)
-
-		// Notify the gallery immediately for every announced path, then fetch.
-		for _, path := range ann.Paths {
-			localPath := d.localDestination(path)
-			result, _ := json.Marshal(map[string]string{"path": localPath})
-			d.broadcast(socketMsg{Type: "image_imported", Data: result})
-		}
-		for _, path := range ann.Paths {
-			localPath := d.localDestination(path)
-			if _, err := os.Stat(localPath + ".proxy.avif"); err == nil {
-				continue // already have it
-			}
-			log.Printf("[proxy] queuing fetch: '%s'", filepath.Base(path))
-			go d.autoFetchProxy(path, ann.BaseURL)
-		}
-	}
-}
-
-// ---------------------------------------------------------------------------
 // Proxy fetching
 // ---------------------------------------------------------------------------
 
@@ -1417,117 +842,6 @@ func (d *daemon) localDestination(remotePath string) string {
 	return remotePath // last resort: will attempt to create dirs
 }
 
-// downloadProxy fetches a proxy from baseURL, writes it to localPath+".proxy.avif",
-// and returns true on success.
-func (d *daemon) downloadProxy(remotePath, localPath, baseURL string) bool {
-	url := baseURL + "/proxy?path=" + remotePath
-	resp, err := d.httpGet(url)
-	if err != nil {
-		log.Printf("[proxy] fetch '%s' from %s: %v", filepath.Base(remotePath), baseURL, err)
-		return false
-	}
-	if resp.StatusCode != 200 {
-		resp.Body.Close()
-		log.Printf("[proxy] fetch '%s' from %s: HTTP %d", filepath.Base(remotePath), baseURL, resp.StatusCode)
-		return false
-	}
-	data, err := io.ReadAll(resp.Body)
-	resp.Body.Close()
-	if err != nil {
-		return false
-	}
-
-	dst := localPath + ".proxy.avif"
-	if err := os.MkdirAll(filepath.Dir(dst), 0755); err != nil {
-		log.Printf("[proxy] mkdir '%s': %v", filepath.Dir(dst), err)
-		return false
-	}
-	if err := os.WriteFile(dst, data, 0644); err != nil {
-		log.Printf("[proxy] write '%s': %v", dst, err)
-		return false
-	}
-	log.Printf("[proxy] fetched '%s' → '%s' (%d KB)", filepath.Base(remotePath), dst, len(data)/1024)
-	return true
-}
-
-// fetchXMPSidecar downloads the XMP sidecar for remotePath from baseURL and
-// writes it to localPath+".xmp". Called after proxy download so mobile starts
-// edits from the full darktable XMP (preserving history and existing rating)
-// rather than from an empty skeleton. Skips silently if the server has no XMP.
-func (d *daemon) fetchXMPSidecar(remotePath, localPath, baseURL string) {
-	xmpURL := baseURL + "/xmp?path=" + url.QueryEscape(remotePath)
-	resp, err := d.httpGet(xmpURL)
-	if err != nil || resp.StatusCode != 200 {
-		if resp != nil {
-			resp.Body.Close()
-		}
-		return
-	}
-	data, err := io.ReadAll(resp.Body)
-	resp.Body.Close()
-	if err != nil || len(data) == 0 {
-		return
-	}
-	dst := localPath + ".xmp"
-	if err := os.WriteFile(dst, data, 0644); err != nil {
-		log.Printf("[xmp] write sidecar '%s': %v", filepath.Base(localPath), err)
-	}
-}
-
-// mergeXMPMeta returns dst with xmp:Rating and darktable:colorlabels patched
-// from src. Used when src is a thin mobile-edit push (no darktable history)
-// and dst is the full desktop XMP so the history stack is not overwritten.
-func mergeXMPMeta(dst, src string) string {
-	reAttr := regexp.MustCompile(`\bxmp:Rating\s*=\s*"[^"]*"`)
-	reElem := regexp.MustCompile(`<xmp:Rating>\s*(\d+)\s*</xmp:Rating>`)
-	reColor := regexp.MustCompile(`(?s)<darktable:colorlabels>.*?</darktable:colorlabels>`)
-
-	result := dst
-
-	// Extract the rating value from src — prefer attribute form, fall back to element.
-	ratingVal := ""
-	if m := reAttr.FindString(src); m != "" {
-		// m is e.g. `xmp:Rating="3"` — extract just the number
-		inner := regexp.MustCompile(`"([^"]*)"`)
-		if sub := inner.FindStringSubmatch(m); len(sub) == 2 {
-			ratingVal = sub[1]
-		}
-	} else if m := reElem.FindStringSubmatch(src); len(m) == 2 {
-		ratingVal = m[1]
-	}
-
-	if ratingVal != "" {
-		// Apply to dst: update attribute form if present, else element form, else insert.
-		if reAttr.MatchString(result) {
-			result = reAttr.ReplaceAllLiteralString(result, `xmp:Rating="`+ratingVal+`"`)
-		} else if reElem.MatchString(result) {
-			result = reElem.ReplaceAllLiteralString(result, "<xmp:Rating>"+ratingVal+"</xmp:Rating>")
-		} else {
-			pos := strings.Index(result, "</rdf:Description>")
-			if pos >= 0 {
-				result = result[:pos] + "  <xmp:Rating>" + ratingVal + "</xmp:Rating>\n" + result[pos:]
-			}
-		}
-	}
-
-	// Patch colorlabels block.
-	if colorSrc := reColor.FindString(src); colorSrc != "" {
-		if reColor.MatchString(result) {
-			result = reColor.ReplaceAllLiteralString(result, colorSrc)
-		}
-	}
-	return result
-}
-
-// extractTagValue returns the text content of an XML element like <Tag>value</Tag>.
-func extractTagValue(elem string) string {
-	re := regexp.MustCompile(`>([^<]*)<`)
-	if m := re.FindStringSubmatch(elem); len(m) == 2 {
-		return strings.TrimSpace(m[1])
-	}
-	return ""
-}
-
 // allPeerURLs returns base URLs of peers that have recently responded to a
 // manifest sync (the periodic "hello").  Only entries in d.syncedPeers are
 // returned — peers that failed their last sync, or that have been discovered
@@ -1543,425 +857,6 @@ func (d *daemon) allPeerURLs() []string {
 		}
 	}
 	return urls
-}
-
-// fetchPreviewFromPeers tries every known peer until it obtains a JPEG preview
-// for canonicalPath at the requested size, then caches it locally.
-// Retries up to 3 additional times with a short delay so that a preview that
-// is still being generated on the desktop side (common right after an edit)
-// will eventually arrive without manual intervention.
-func (d *daemon) fetchPreviewFromPeers(canonicalPath, size string) {
-	localPath := d.localDestination(canonicalPath)
-
-	log.Printf("[preview] fetching '%s' size=%s from peers",
-		filepath.Base(canonicalPath), size)
-
-	// On mobile, canonicalPath is the local import-dir path which differs from
-	// the desktop's original path.  Use localToRemote to get the path the
-	// desktop knows about (same logic fetchProxyFromPeer uses).
-	remotePath := canonicalPath
-	d.localToRemoteMu.RLock()
-	if r, ok := d.localToRemote[canonicalPath]; ok {
-		remotePath = r
-	}
-	d.localToRemoteMu.RUnlock()
-
-	dst := localPath + ".preview-" + size + ".jpg"
-
-	const maxAttempts = 8
-	const retryDelay = 4 * time.Second
-
-	for attempt := 0; attempt < maxAttempts; attempt++ {
-		if attempt > 0 {
-			time.Sleep(retryDelay)
-		}
-		for _, baseURL := range d.allPeerURLs() {
-			d.fetchPreviewJPEG(remotePath, localPath, baseURL, size)
-			if fileExists(dst) {
-				return
-			}
-		}
-	}
-}
-
-// autoFetchProxy downloads a proxy and notifies darktable to import it.
-func (d *daemon) autoFetchProxy(remotePath, baseURL string) {
-	localPath := d.localDestination(remotePath)
-
-	// On the phone, localPath differs from remotePath (importDir vs desktop path).
-	// Remember the mapping so fetch_proxy and fetch_preview can resolve it back.
-	if localPath != remotePath {
-		d.localToRemoteMu.Lock()
-		d.localToRemote[localPath] = remotePath
-		d.localToRemoteMu.Unlock()
-		d.markPathMapDirty()
-	}
-
-	// Fetch the thumbnail JPEG first — it's small and lets the mobile gallery
-	// show a preview immediately, even while the larger AVIF is still downloading.
-	d.fetchPreviewJPEG(remotePath, localPath, baseURL, "thumb")
-
-	if !d.downloadProxy(remotePath, localPath, baseURL) {
-		return
-	}
-
-	// Download the XMP sidecar so mobile always starts edits from the full
-	// darktable XMP (preserving history stack and existing rating).
-	d.fetchXMPSidecar(remotePath, localPath, baseURL)
-
-	d.addToLocalIndex(localPath)
-
-	// Re-broadcast so darktable (desktop) imports the proxy path and the mobile
-	// model updates hasProxy = true once the AVIF is locally present.
-	result, _ := json.Marshal(map[string]string{"path": localPath})
-	d.broadcast(socketMsg{Type: "image_imported", Data: result})
-}
-
-// fetchPreviewJPEG downloads a JPEG preview from a peer and caches it beside
-// the local proxy AVIF. size must be "thumb" or "full".
-//
-// Uses a conditional GET (If-Modified-Since) when a cached copy already exists,
-// so the file is only re-downloaded when the desktop has a fresher version.
-// Broadcasts preview_updated when an existing preview is replaced.
-func (d *daemon) fetchPreviewJPEG(remotePath, localPath, baseURL, size string) {
-	dst := localPath + ".preview-" + size + ".jpg"
-	previewURL := baseURL + "/preview?path=" + url.QueryEscape(remotePath) + "&size=" + size
-
-	req, err := http.NewRequest(http.MethodGet, previewURL, nil)
-	if err != nil {
-		return
-	}
-	req.Header.Set(passphraseHeader, d.authToken)
-
-	// Conditional GET: ask the server to skip the body if nothing changed.
-	hadCached := false
-	if fi, err := os.Stat(dst); err == nil {
-		hadCached = true
-		req.Header.Set("If-Modified-Since", fi.ModTime().UTC().Format(http.TimeFormat))
-	}
-
-	resp, err := d.tlsClient.Do(req)
-	if err != nil {
-		log.Printf("[preview] fetch %s from %s: %v", filepath.Base(remotePath), baseURL, err)
-		return
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode == http.StatusNotModified {
-		return // our copy is still current
-	}
-	if resp.StatusCode != http.StatusOK {
-		log.Printf("[preview] fetch %s from %s: HTTP %d", filepath.Base(remotePath), baseURL, resp.StatusCode)
-		return
-	}
-
-	data, err := io.ReadAll(resp.Body)
-	if err != nil || len(data) == 0 {
-		return
-	}
-
-	if err := os.WriteFile(dst, data, 0644); err != nil {
-		return
-	}
-
-	verb := "fetched"
-	if hadCached {
-		verb = "refreshed"
-	}
-	log.Printf("[preview] %s %s preview for '%s' (%d KB)", verb, size, filepath.Base(remotePath), len(data)/1024)
-	result, _ := json.Marshal(map[string]string{"path": localPath})
-	d.broadcast(socketMsg{Type: "preview_updated", Data: result})
-}
-
-// fetchProxyFromPeer handles an explicit "fetch_proxy" command from darktable.
-func (d *daemon) fetchProxyFromPeer(localPath string, enc *json.Encoder) {
-	// On the phone, the command arrives with the local importDir path.
-	// Resolve it back to the peer's canonical path for the HTTP request,
-	// but always write the result to the local path.
-	requestPath := localPath
-	d.localToRemoteMu.RLock()
-	if remote, ok := d.localToRemote[localPath]; ok {
-		requestPath = remote
-	}
-	d.localToRemoteMu.RUnlock()
-
-	for _, baseURL := range d.allPeerURLs() {
-		fetchURL := baseURL + "/proxy?path=" + url.QueryEscape(requestPath)
-		resp, err := d.httpGet(fetchURL)
-		if err != nil || resp.StatusCode != 200 {
-			if resp != nil {
-				resp.Body.Close()
-			}
-			continue
-		}
-		data, err := io.ReadAll(resp.Body)
-		resp.Body.Close()
-		if err != nil {
-			continue
-		}
-		dst := localPath + ".proxy.avif"
-		if err := os.MkdirAll(filepath.Dir(dst), 0755); err == nil {
-			if err := os.WriteFile(dst, data, 0644); err == nil {
-				log.Printf("[proxy] explicit fetch '%s' from %s (%d KB)",
-					filepath.Base(localPath), baseURL, len(data)/1024)
-				result, _ := json.Marshal(map[string]string{"path": dst, "status": "ok"})
-				enc.Encode(socketMsg{Type: "proxy_fetched", Data: result})
-				return
-			}
-		}
-	}
-	result, _ := json.Marshal(map[string]string{"path": localPath, "status": "not_found"})
-	enc.Encode(socketMsg{Type: "proxy_fetched", Data: result})
-}
-
-// ---------------------------------------------------------------------------
-// Proxy announce
-// ---------------------------------------------------------------------------
-
-func (d *daemon) announceProxies() {
-	tick := time.NewTicker(30 * time.Second)
-	defer tick.Stop()
-	d.publishProxyAnnounce()
-	for {
-		select {
-		case <-d.ctx.Done():
-			return
-		case <-tick.C:
-			d.publishProxyAnnounce()
-		}
-	}
-}
-
-func (d *daemon) publishProxyAnnounce() {
-	d.manifestMu.Lock()
-	if d.manifestHash == "" {
-		d.rebuildManifestLocked()
-	}
-	paths := d.manifestPaths
-	d.manifestMu.Unlock()
-
-	ann := proxyAnnounce{
-		SenderID: d.host.ID().String(),
-		BaseURL:  d.localProxyURL(),
-		Paths:    paths,
-	}
-	if b, err := json.Marshal(ann); err == nil {
-		d.proxyTop.Publish(d.ctx, b)
-		if len(paths) > 0 {
-			log.Printf("[announce] broadcasting %d proxies", len(paths))
-		}
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Static peer sync
-// ---------------------------------------------------------------------------
-
-func (d *daemon) syncStaticPeers() {
-	time.Sleep(2 * time.Second)
-	for _, u := range d.peersToSync() {
-		go d.syncWithPeer(u)
-	}
-	tick := time.NewTicker(60 * time.Second)
-	defer tick.Stop()
-	for {
-		select {
-		case <-d.ctx.Done():
-			return
-		case <-tick.C:
-			for _, u := range d.peersToSync() {
-				go d.syncWithPeer(u)
-			}
-		}
-	}
-}
-
-// peersToSync returns the union of static peers and DB-persisted peers,
-// deduplicated.
-func (d *daemon) peersToSync() []string {
-	seen := make(map[string]bool)
-	var out []string
-	add := func(u string) {
-		if u != "" && !seen[u] {
-			seen[u] = true
-			out = append(out, u)
-		}
-	}
-	for _, u := range d.staticPeers {
-		add(u)
-	}
-	for _, u := range d.pdb.allURLs() {
-		add(u)
-	}
-	return out
-}
-
-func (d *daemon) syncWithPeer(baseURL string) {
-	baseURL = strings.TrimRight(baseURL, "/")
-	if baseURL == "" {
-		return
-	}
-
-	// Never connect to ourselves.
-	myURL := d.localProxyURL()
-	if baseURL == myURL || baseURL == fmt.Sprintf("https://%s:%d", d.localIP(), proxyHTTPPort) {
-		return
-	}
-
-	// Check first-visit state but don't commit until we know the HTTP call
-	// succeeds — otherwise a failed attempt marks firstVisit=false forever and
-	// the libp2p connect + reverse-peer handshake never happens.
-	d.syncedPeersMu.Lock()
-	firstVisit := !d.syncedPeers[baseURL]
-	d.syncedPeersMu.Unlock()
-
-	manifestURL := baseURL + "/manifest"
-	if firstVisit {
-		manifestURL += "?from=" + url.QueryEscape(myURL)
-	}
-
-	resp, err := d.httpGet(manifestURL)
-	if err != nil {
-		log.Printf("[peer] manifest from %s: %v", baseURL, err)
-		d.pdb.markFailure(baseURL)
-		// Reset so the next successful attempt still performs the firstVisit steps.
-		d.syncedPeersMu.Lock()
-		delete(d.syncedPeers, baseURL)
-		d.syncedPeersMu.Unlock()
-		// If the failure is a fingerprint rejection, probe the peer's cert so
-		// we can present it to the user for acceptance in preferences.
-		if strings.Contains(err.Error(), "not in allowed list") {
-			go func(u string) {
-				if fp := probePeerFingerprint(u); fp != "" {
-					d.recordCandidatePeer(u, fp)
-				}
-			}(baseURL)
-		}
-		return
-	}
-	if resp.StatusCode != 200 {
-		resp.Body.Close()
-		log.Printf("[peer] %s: HTTP %d (old daemon?)", baseURL, resp.StatusCode)
-		d.pdb.markFailure(baseURL)
-		d.syncedPeersMu.Lock()
-		delete(d.syncedPeers, baseURL)
-		d.syncedPeersMu.Unlock()
-		return
-	}
-
-	// Commit first-visit only after confirmed success, with a CAS to handle
-	// concurrent goroutines both racing to be the first successful contact.
-	d.syncedPeersMu.Lock()
-	firstVisit = !d.syncedPeers[baseURL]
-	d.syncedPeers[baseURL] = true
-	d.syncedPeersMu.Unlock()
-	var m manifestResp
-	if err := json.NewDecoder(resp.Body).Decode(&m); err != nil {
-		resp.Body.Close()
-		log.Printf("[peer] manifest decode from %s: %v", baseURL, err)
-		return
-	}
-	resp.Body.Close()
-
-	// Successful contact — record in DB.
-	d.pdb.markSuccess(baseURL, m.PeerID)
-
-	peerShort := m.PeerID
-	if len(peerShort) > 12 {
-		peerShort = peerShort[:12]
-	}
-
-	// Skip the expensive per-path broadcast + stat work when the remote's
-	// manifest content hasn't changed since our last successful sync.
-	// (Hash is non-empty only on daemons that support this field.)
-	d.peerHashMu.Lock()
-	sameHash := m.Hash != "" && m.Hash == d.peerLastHash[baseURL]
-	if m.Hash != "" {
-		d.peerLastHash[baseURL] = m.Hash
-	}
-	d.peerHashMu.Unlock()
-
-	if sameHash && !firstVisit {
-		log.Printf("[peer] %s: manifest unchanged (hash=%s), skipping per-path work (%d proxies)",
-			baseURL, m.Hash, len(m.Paths))
-		return
-	}
-
-	log.Printf("[peer] %s: peerID=%s  proxies=%d  indirect=%d",
-		baseURL, peerShort, len(m.Paths), len(m.KnownPeers))
-
-	// Connect via libp2p so gossipsub (XMP) works (once per peer).
-	if firstVisit && m.PeerID != "" && m.LibP2PIP != "" && m.LibP2PPort > 0 {
-		maStr := fmt.Sprintf("/ip4/%s/tcp/%d", m.LibP2PIP, m.LibP2PPort)
-		ma, maErr := multiaddr.NewMultiaddr(maStr)
-		pid, pidErr := peer.Decode(m.PeerID)
-		if maErr == nil && pidErr == nil && pid != d.host.ID() {
-			pinfo := peer.AddrInfo{ID: pid, Addrs: []multiaddr.Multiaddr{ma}}
-			d.peersMu.Lock()
-			d.peers[pid] = baseURL
-			d.peersMu.Unlock()
-			if err := d.host.Connect(d.ctx, pinfo); err != nil {
-				log.Printf("[peer] libp2p connect to %s: %v", maStr, err)
-			} else {
-				log.Printf("[peer] libp2p connected to %s", peerShort)
-			}
-		}
-	}
-
-	// Immediately notify subscribers of every image the peer knows about so the
-	// mobile gallery populates at once without waiting for AVIF downloads.
-	// image_imported deduplicates on the receiver side, so re-announcing on
-	// every sync is harmless.
-	for _, path := range m.Paths {
-		localPath := d.localDestination(path)
-		result, _ := json.Marshal(map[string]string{"path": localPath})
-		d.broadcast(socketMsg{Type: "image_imported", Data: result})
-	}
-
-	// Now start background downloads: thumbnail JPEG (fast) then AVIF proxy.
-	// downloadSem caps concurrency so a large library doesn't spawn thousands
-	// of goroutines simultaneously.
-	missing := 0
-	for _, path := range m.Paths {
-		localPath := d.localDestination(path)
-		if _, err := os.Stat(localPath + ".proxy.avif"); err == nil {
-			// AVIF already present; just refresh the thumbnail if it changed.
-			go func(p, lp, bu string) {
-				d.downloadSem <- struct{}{}
-				defer func() { <-d.downloadSem }()
-				d.fetchPreviewJPEG(p, lp, bu, "thumb")
-			}(path, localPath, baseURL)
-			continue
-		}
-		missing++
-		go func(p, bu string) {
-			d.downloadSem <- struct{}{}
-			defer func() { <-d.downloadSem }()
-			d.autoFetchProxy(p, bu)
-		}(path, baseURL)
-	}
-	if missing > 0 {
-		log.Printf("[peer] fetching %d missing proxies from %s", missing, baseURL)
-	} else if len(m.Paths) > 0 {
-		log.Printf("[peer] all %d proxies from %s already present", len(m.Paths), baseURL)
-	}
-
-	// Gossip: connect to indirect peers the remote knows about (first visit only
-	// to avoid propagation storms when the mesh is fully connected).
-	if firstVisit {
-		for _, peerURL := range m.KnownPeers {
-			if peerURL != "" && peerURL != myURL {
-				d.pdb.touch(peerURL, "") // persist before we try to reach them
-				go d.syncWithPeer(peerURL)
-			}
-		}
-		// If the remote advertises a different external URL, prefer that for future calls.
-		if m.ExternalURL != "" && m.ExternalURL != baseURL {
-			d.syncedPeersMu.Lock()
-			d.syncedPeers[m.ExternalURL] = true // treat as same peer, avoid double-sync
-			d.syncedPeersMu.Unlock()
-		}
-	}
 }
 
 // ---------------------------------------------------------------------------
@@ -2134,55 +1029,6 @@ func (d *daemon) isAnnouncedPath(rawPath string) bool {
 	return err == nil
 }
 
-func (d *daemon) serveXMP(w http.ResponseWriter, r *http.Request) {
-	if r.Method == http.MethodGet {
-		// GET /xmp?path=<canonical> — serve the XMP sidecar for an announced path.
-		canonicalPath := r.URL.Query().Get("path")
-		if canonicalPath == "" || strings.Contains(canonicalPath, "..") {
-			http.Error(w, "bad request", 400)
-			return
-		}
-		if !d.isAnnouncedPath(canonicalPath) {
-			http.Error(w, "not found", 404)
-			return
-		}
-		xmpPath := canonicalPath + ".xmp"
-		data, err := os.ReadFile(xmpPath)
-		if err != nil {
-			http.Error(w, "not found", 404)
-			return
-		}
-		w.Header().Set("Content-Type", "application/xml")
-		w.Write(data)
-		return
-	}
-	if r.Method != http.MethodPost {
-		http.Error(w, "method not allowed", http.StatusMethodNotAllowed)
-		return
-	}
-	var x xmpMsg
-	if err := json.NewDecoder(r.Body).Decode(&x); err != nil {
-		http.Error(w, "bad request", http.StatusBadRequest)
-		return
-	}
-	w.WriteHeader(http.StatusNoContent)
-	go d.applyInboundXMP(x, "http:"+r.RemoteAddr)
-}
-
-func (d *daemon) serveProxy(w http.ResponseWriter, r *http.Request) {
-	canonicalPath := r.URL.Query().Get("path")
-	if canonicalPath == "" || strings.Contains(canonicalPath, "..") {
-		http.Error(w, "bad request", 400)
-		return
-	}
-	if !d.isAnnouncedPath(canonicalPath) {
-		http.Error(w, "not found", 404)
-		return
-	}
-	log.Printf("[http] serve proxy '%s' to %s", filepath.Base(canonicalPath), r.RemoteAddr)
-	http.ServeFile(w, r, canonicalPath+".proxy.avif")
-}
-
 // rebuildManifestLocked rebuilds d.manifestPaths and d.manifestHash.
 // Must be called with d.manifestMu held.
 func (d *daemon) rebuildManifestLocked() {
@@ -2247,438 +1093,6 @@ func (d *daemon) serveManifest(w http.ResponseWriter, r *http.Request) {
 	json.NewEncoder(w).Encode(resp)
 }
 
-// ---------------------------------------------------------------------------
-// JPEG preview endpoint
-// ---------------------------------------------------------------------------
-
-// servePreview serves JPEG previews for raw images.
-//
-// GET /preview?path=<canonical-raw-path>&size=thumb|full
-//
-// Sources, in priority order:
-//  1. Cached preview JPEG beside the proxy AVIF (.preview-thumb.jpg / .preview-full.jpg)
-//  2. darktable's mipmap cache (~/.cache/darktable/mipmaps-*.d/)
-//  3. darktable-cli export (desktop only; slow but guaranteed)
-//
-// size=thumb picks mipmap sizes [1,2] (~337–450 px wide).
-// size=full  picks mipmap sizes [4,6,3] (~1800/3800/1350 px wide).
-//
-// Uses http.ServeContent so clients get proper Last-Modified and 304 support.
-func (d *daemon) servePreview(w http.ResponseWriter, r *http.Request) {
-	rawPath := r.URL.Query().Get("path")
-	if rawPath == "" || strings.Contains(rawPath, "..") {
-		http.Error(w, "bad request", http.StatusBadRequest)
-		return
-	}
-	if !d.isAnnouncedPath(rawPath) {
-		http.Error(w, "not found", http.StatusNotFound)
-		return
-	}
-
-	sizeParam := r.URL.Query().Get("size")
-	var (
-		cacheSuffix string
-		mipmapSizes []int
-		maxDim      int
-	)
-	if sizeParam == "full" {
-		cacheSuffix = ".preview-full.jpg"
-		mipmapSizes = []int{4, 6, 3}
-		maxDim = 1920
-	} else {
-		cacheSuffix = ".preview-thumb.jpg"
-		mipmapSizes = []int{1, 2}
-		maxDim = 480
-	}
-
-	cachePath := rawPath + cacheSuffix
-
-	// Delete a stale cache file that lacks an ICC profile so it is regenerated
-	// with the correct color metadata on this request.
-	if fileExists(cachePath) {
-		if existing, err := os.ReadFile(cachePath); err == nil && !hasJPEGICCProfile(existing) {
-			os.Remove(cachePath)
-			log.Printf("[preview] removed ICC-less cache for '%s' size=%s; regenerating",
-				filepath.Base(rawPath), sizeParam)
-		}
-	}
-
-	// Populate the cache file if it doesn't exist yet.
-	if !fileExists(cachePath) {
-		mipmapData := d.findMipmapJPEG(rawPath, mipmapSizes)
-
-		// For full-size previews, mipmap cache JEPGs must go through
-		// darktable-cli to guarantee ICC profile + full Exif metadata (camera
-		// make/model/lens/exposure/GPS) so sharing to other apps is complete.
-		if sizeParam == "full" && mipmapData != nil {
-			log.Printf("[preview] full size: skipping mipmap for '%s'; using darktable-cli for ICC+Exif",
-				filepath.Base(rawPath))
-			mipmapData = nil
-		}
-
-		if mipmapData != nil {
-			// Thumbnails: mipmap JPEG lacks an ICC marker. Inject sRGB so mobile
-			// apps render consistently with the full preview (also sRGB). Using
-			// the desktop display ICC here would make the gallery look vivid but
-			// sharing would appear muted by comparison in non-colour-managed apps.
-			mipmapData = injectICCProfile(mipmapData, compactSRGBICC())
-			if err := os.WriteFile(cachePath, mipmapData, 0644); err == nil {
-				log.Printf("[preview] mipmap+sRGB-ICC → cache '%s' size=%s (%d KB)",
-					filepath.Base(rawPath), sizeParam, len(mipmapData)/1024)
-			}
-		} else {
-			// Skip darktable-cli when the GUI is running — it would open its own
-			// database connection and race with the GUI's writers causing
-			// SQLITE_BUSY assertions.  The GUI regenerates its mipmap cache on
-			// its own after an XMP reload; mobile can retry in a few seconds.
-			d.subsMu.Lock()
-			darktableRunning := len(d.subs) > 0
-			d.subsMu.Unlock()
-			if darktableRunning {
-				// Ask darktable to generate the mipmap in the background.
-				// It will write the file to its mipmap cache; the retry loop
-				// in fetchPreviewFromPeers will pick it up a few seconds later.
-				pathJSON, _ := json.Marshal(rawPath)
-				d.broadcast(socketMsg{
-					Type: "generate_preview",
-					Data: json.RawMessage(`{"path":` + string(pathJSON) + `}`),
-				})
-				log.Printf("[preview] requested mipmap generation for '%s' from darktable",
-					filepath.Base(rawPath))
-				http.Error(w, "preview regenerating", http.StatusServiceUnavailable)
-				return
-			}
-			if _, err := exportWithDarktableCLI(rawPath, cachePath, maxDim); err != nil {
-				http.Error(w, "preview not available", http.StatusNotFound)
-				return
-			}
-			log.Printf("[preview] cli export '%s' size=%s", filepath.Base(rawPath), sizeParam)
-		}
-	}
-
-	// Serve the cache file. http.ServeContent handles Last-Modified and 304.
-	f, err := os.Open(cachePath)
-	if err != nil {
-		http.Error(w, "preview not available", http.StatusNotFound)
-		return
-	}
-	defer f.Close()
-	fi, err := f.Stat()
-	if err != nil {
-		http.Error(w, "stat failed", http.StatusInternalServerError)
-		return
-	}
-	w.Header().Set("Content-Type", "image/jpeg")
-	http.ServeContent(w, r, "", fi.ModTime(), f)
-}
-
-// findMipmapJPEG looks up the image ID in darktable's library.db and returns
-// the first mipmap JPEG found for the given size preference list.
-func (d *daemon) findMipmapJPEG(rawPath string, sizes []int) []byte {
-	mipmapDir, err := darktableMipmapDir()
-	if err != nil {
-		return nil
-	}
-	libPath := darktableLibraryPath()
-	if libPath == "" {
-		return nil
-	}
-	imgID, err := lookupDarktableImageID(libPath, rawPath)
-	if err != nil || imgID == 0 {
-		return nil
-	}
-	for _, sz := range sizes {
-		p := filepath.Join(mipmapDir, fmt.Sprintf("%d", sz), fmt.Sprintf("%d.jpg", imgID))
-		if data, err := os.ReadFile(p); err == nil {
-			return data
-		}
-	}
-	return nil
-}
-
-// darktableLibraryPath returns the path to darktable's library.db, or "".
-func darktableLibraryPath() string {
-	cfgDir, err := os.UserConfigDir()
-	if err != nil {
-		return ""
-	}
-	p := filepath.Join(cfgDir, "darktable", "library.db")
-	if _, err := os.Stat(p); err == nil {
-		return p
-	}
-	return ""
-}
-
-// darktableMipmapDir returns the path to the darktable mipmap cache directory.
-func darktableMipmapDir() (string, error) {
-	cacheDir, err := os.UserCacheDir()
-	if err != nil {
-		return "", err
-	}
-	dtCache := filepath.Join(cacheDir, "darktable")
-	entries, err := os.ReadDir(dtCache)
-	if err != nil {
-		return "", err
-	}
-	for _, e := range entries {
-		if e.IsDir() && strings.HasPrefix(e.Name(), "mipmaps-") && strings.HasSuffix(e.Name(), ".d") {
-			return filepath.Join(dtCache, e.Name()), nil
-		}
-	}
-	return "", fmt.Errorf("no mipmaps directory in %s", dtCache)
-}
-
-// lookupDarktableImageID queries library.db for the image ID matching rawPath.
-func lookupDarktableImageID(libPath, rawPath string) (int64, error) {
-	// Open read-only; WAL mode so we don't interfere with a running darktable.
-	db, err := sql.Open("sqlite", libPath+"?mode=ro&_journal=WAL")
-	if err != nil {
-		return 0, err
-	}
-	defer db.Close()
-
-	dir := filepath.Dir(rawPath)
-	name := filepath.Base(rawPath)
-	var id int64
-	err = db.QueryRow(`
-		SELECT img.id
-		FROM images img
-		JOIN film_rolls f ON img.film_id = f.id
-		WHERE img.filename = ? AND f.folder = ?`,
-		name, dir).Scan(&id)
-	if err != nil {
-		return 0, err
-	}
-	return id, nil
-}
-
-// exportWithDarktableCLI renders rawPath to outPath via darktable-cli and
-// returns the JPEG bytes. Applies the XMP sidecar if present.
-//
-// Produces sRGB output with relative-colorimetric rendering intent (the
-// standard choice for photographic output), then copies the full Exif block
-// from the raw file into the JPEG using exiftool so the recipient has all
-// camera metadata (make, model, lens, exposure, GPS, etc.).
-func exportWithDarktableCLI(rawPath, outPath string, maxDim int) ([]byte, error) {
-	args := []string{rawPath}
-	if xmp := rawPath + ".xmp"; fileExists(xmp) {
-		args = append(args, xmp)
-	}
-	args = append(args, outPath,
-		"--width", fmt.Sprintf("%d", maxDim),
-		"--height", fmt.Sprintf("%d", maxDim),
-		"--icc-type", "SRGB",
-		"--icc-intent", "RELATIVE_COLORIMETRIC",
-	)
-	cmd := exec.Command("darktable-cli", args...)
-	if out, err := cmd.CombinedOutput(); err != nil {
-		return nil, fmt.Errorf("darktable-cli: %w\n%s", err, out)
-	}
-
-	// Copy all Exif tags from the raw source into the rendered JPEG.
-	// darktable-cli strips Exif from exports; exiftool restores it.
-	// -overwrite_original avoids leaving a _original backup file.
-	// -tagsFromFile copies all tags; -m ignores minor errors.
-	if exiftoolPath, err := exec.LookPath("exiftool"); err == nil {
-		exif := exec.Command(exiftoolPath,
-			"-overwrite_original",
-			"-m",
-			"-tagsFromFile", rawPath,
-			"-Exif:All", // all Exif tags (camera make/model/lens/exposure/GPS…)
-			"-IPTC:All", // keywords, caption, copyright
-			outPath,
-		)
-		if out, err := exif.CombinedOutput(); err != nil {
-			log.Printf("[preview] exiftool copy from %s: %v\n%s", filepath.Base(rawPath), err, out)
-			// Non-fatal: JPEG is still usable, just lacks full Exif.
-		}
-	} else {
-		log.Printf("[preview] exiftool not found; skipping Exif copy for %s", filepath.Base(rawPath))
-	}
-
-	return os.ReadFile(outPath)
-}
-
-// loadDisplayICC returns the bytes of the desktop monitor's ICC profile.
-// Resolution order:
-//  1. Explicit path from --display-icc flag.
-//  2. Active display profile from colord (colormgr CLI).
-//  3. First 'mntr'-class ICC file found in ~/.local/share/icc/.
-//  4. Hardcoded compact sRGB profile as a universal fallback.
-func loadDisplayICC(explicitPath string) []byte {
-	if explicitPath != "" {
-		if data, err := os.ReadFile(explicitPath); err == nil {
-			log.Printf("[color] display ICC from --display-icc (%d bytes)", len(data))
-			return data
-		} else {
-			log.Printf("[color] --display-icc %s: %v; trying auto-detect", explicitPath, err)
-		}
-	}
-
-	if p := findDisplayICCViaColormgr(); p != "" {
-		if data, err := os.ReadFile(p); err == nil {
-			log.Printf("[color] display ICC from colord: %s (%d bytes)", filepath.Base(p), len(data))
-			return data
-		}
-	}
-
-	if home, err := os.UserHomeDir(); err == nil {
-		iccDir := filepath.Join(home, ".local", "share", "icc")
-		entries, _ := os.ReadDir(iccDir)
-		for _, e := range entries {
-			if !strings.HasSuffix(e.Name(), ".icc") {
-				continue
-			}
-			p := filepath.Join(iccDir, e.Name())
-			data, err := os.ReadFile(p)
-			if err != nil || len(data) < 16 {
-				continue
-			}
-			if string(data[12:16]) == "mntr" { // ICC profile class = display
-				log.Printf("[color] display ICC from %s (%d bytes)", e.Name(), len(data))
-				return data
-			}
-		}
-	}
-
-	log.Printf("[color] display ICC not detected; using compact sRGB fallback")
-	return compactSRGBICC()
-}
-
-// findDisplayICCViaColormgr asks colord for the active display profile path.
-func findDisplayICCViaColormgr() string {
-	out, err := exec.Command("colormgr", "get-devices").Output()
-	if err != nil {
-		return ""
-	}
-	for _, line := range strings.Split(string(out), "\n") {
-		line = strings.TrimSpace(line)
-		if !strings.HasPrefix(line, "Object Path:") || !strings.Contains(line, "display") {
-			continue
-		}
-		devicePath := strings.TrimSpace(strings.TrimPrefix(line, "Object Path:"))
-		prof, err := exec.Command("colormgr", "device-get-default-profile", devicePath).Output()
-		if err != nil {
-			continue
-		}
-		for _, pl := range strings.Split(string(prof), "\n") {
-			pl = strings.TrimSpace(pl)
-			if strings.HasPrefix(pl, "Filename:") {
-				return strings.TrimSpace(strings.TrimPrefix(pl, "Filename:"))
-			}
-		}
-	}
-	return ""
-}
-
-// injectICCProfile inserts icc as an APP2 ICC_PROFILE segment immediately
-// after the JPEG SOI marker. Returns data unchanged if it already has a profile.
-func injectICCProfile(data, icc []byte) []byte {
-	if len(data) < 2 || data[0] != 0xFF || data[1] != 0xD8 {
-		return data
-	}
-	if hasJPEGICCProfile(data) {
-		return data
-	}
-	const iccHeader = "ICC_PROFILE\x00\x01\x01" // single chunk
-	segPayload := append([]byte(iccHeader), icc...)
-	segLen := len(segPayload) + 2
-	app2 := []byte{0xFF, 0xE2, byte(segLen >> 8), byte(segLen)}
-	app2 = append(app2, segPayload...)
-	result := make([]byte, 0, len(data)+len(app2))
-	result = append(result, data[:2]...)
-	result = append(result, app2...)
-	result = append(result, data[2:]...)
-	return result
-}
-
-// injectDisplayICC embeds the daemon's display ICC into a JPEG (for thumbnails).
-func (d *daemon) injectDisplayICC(data []byte) []byte {
-	return injectICCProfile(data, d.displayICC)
-}
-
-// compactSRGBICC returns a compact IEC 61966-2-1 sRGB ICC v2 profile (~212 bytes)
-// used as a fallback when no system display profile can be detected.
-func compactSRGBICC() []byte {
-	return []byte{
-		0x00, 0x00, 0x00, 0xd4, 0x6c, 0x63, 0x6d, 0x73, 0x02, 0x10, 0x00, 0x00,
-		0x6d, 0x6e, 0x74, 0x72, 0x52, 0x47, 0x42, 0x20, 0x58, 0x59, 0x5a, 0x20,
-		0x07, 0xe7, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-		0x61, 0x63, 0x73, 0x70, 0x4d, 0x53, 0x46, 0x54, 0x00, 0x00, 0x00, 0x00,
-		0x73, 0x61, 0x77, 0x73, 0x63, 0x74, 0x72, 0x6c, 0x00, 0x00, 0x00, 0x00,
-		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xf6, 0xd6,
-		0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0xd3, 0x2d, 0x6c, 0x63, 0x6d, 0x73,
-		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-		0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x09,
-		0x64, 0x65, 0x73, 0x63, 0x00, 0x00, 0x00, 0xf0, 0x00, 0x00, 0x00, 0x12,
-		0x77, 0x74, 0x70, 0x74, 0x00, 0x00, 0x01, 0x04, 0x00, 0x00, 0x00, 0x14,
-		0x72, 0x58, 0x59, 0x5a, 0x00, 0x00, 0x01, 0x18, 0x00, 0x00, 0x00, 0x14,
-		0x67, 0x58, 0x59, 0x5a, 0x00, 0x00, 0x01, 0x2c, 0x00, 0x00, 0x00, 0x14,
-		0x62, 0x58, 0x59, 0x5a, 0x00, 0x00, 0x01, 0x40, 0x00, 0x00, 0x00, 0x14,
-		0x72, 0x54, 0x52, 0x43, 0x00, 0x00, 0x01, 0x54, 0x00, 0x00, 0x00, 0x28,
-		0x67, 0x54, 0x52, 0x43, 0x00, 0x00, 0x01, 0x54, 0x00, 0x00, 0x00, 0x28,
-		0x62, 0x54, 0x52, 0x43, 0x00, 0x00, 0x01, 0x54, 0x00, 0x00, 0x00, 0x28,
-		0x63, 0x70, 0x72, 0x74, 0x00, 0x00, 0x01, 0x7c, 0x00, 0x00, 0x00, 0x58,
-		// desc "sRGB"
-		0x6d, 0x6c, 0x75, 0x63, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01,
-		0x00, 0x00, 0x00, 0x0c, 0x65, 0x6e, 0x55, 0x53, 0x00, 0x00, 0x00, 0x04,
-		0x00, 0x00, 0x00, 0x00, 0x00, 0x73, 0x00, 0x52, 0x00, 0x47, 0x00, 0x42,
-		// wtpt D50
-		0x58, 0x59, 0x5a, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xf3, 0x52,
-		0x00, 0x01, 0x00, 0x00, 0x00, 0x01, 0x16, 0xcc,
-		// rXYZ
-		0x58, 0x59, 0x5a, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x6f, 0xa0,
-		0x00, 0x00, 0x38, 0xf5, 0x00, 0x00, 0x03, 0x90,
-		// gXYZ
-		0x58, 0x59, 0x5a, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x62, 0x97,
-		0x00, 0x00, 0xb7, 0x87, 0x00, 0x00, 0x18, 0xda,
-		// bXYZ
-		0x58, 0x59, 0x5a, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x24, 0xa0,
-		0x00, 0x00, 0x0f, 0x84, 0x00, 0x00, 0xb6, 0xc4,
-		// TRC (parametric curve for sRGB gamma)
-		0x70, 0x61, 0x72, 0x61, 0x00, 0x00, 0x00, 0x00, 0x00, 0x03, 0x00, 0x00,
-		0x00, 0x02, 0x4c, 0x1d, 0x00, 0x0a, 0x92, 0x6b, 0x00, 0x01, 0x4a, 0xb6,
-		0x00, 0x00, 0x40, 0x6e, 0x00, 0x00, 0x09, 0x6c,
-		// cprt
-		0x74, 0x65, 0x78, 0x74, 0x00, 0x00, 0x00, 0x00, 0x43, 0x72, 0x65, 0x61,
-		0x74, 0x65, 0x64, 0x20, 0x62, 0x79, 0x20, 0x6c, 0x63, 0x6d, 0x73, 0x20,
-		0x32, 0x2e, 0x31, 0x30, 0x20, 0x28, 0x63, 0x29, 0x20, 0x32, 0x30, 0x31,
-		0x32, 0x2e, 0x20, 0x49, 0x45, 0x43, 0x20, 0x36, 0x31, 0x39, 0x36, 0x36,
-		0x2d, 0x32, 0x2d, 0x31, 0x20, 0x44, 0x65, 0x66, 0x61, 0x75, 0x6c, 0x74,
-		0x20, 0x52, 0x47, 0x42, 0x20, 0x43, 0x6f, 0x6c, 0x6f, 0x75, 0x72, 0x20,
-		0x53, 0x70, 0x61, 0x63, 0x65, 0x20, 0x2d, 0x20, 0x73, 0x52, 0x47, 0x42,
-		0x00, 0x00, 0x00, 0x00,
-	}
-}
-
-// hasJPEGICCProfile reports whether the JPEG byte slice contains an APP2
-// segment whose payload begins with the "ICC_PROFILE\x00" signature.
-func hasJPEGICCProfile(data []byte) bool {
-	const iccSig = "ICC_PROFILE\x00"
-	// Walk APP markers. JPEG starts with FF D8; APP markers are FF Ex/FF Ex.
-	for i := 2; i+4 < len(data); {
-		if data[i] != 0xFF {
-			break
-		}
-		marker := data[i+1]
-		segLen := int(data[i+2])<<8 | int(data[i+3]) // includes 2-byte length field
-		if segLen < 2 {
-			break
-		}
-		payloadStart := i + 4
-		payloadLen := segLen - 2
-		// APP2 = 0xE2
-		if marker == 0xE2 && payloadLen >= len(iccSig) {
-			if string(data[payloadStart:payloadStart+len(iccSig)]) == iccSig {
-				return true
-			}
-		}
-		i += 2 + segLen
-	}
-	return false
-}
-
 func fileExists(p string) bool {
 	_, err := os.Stat(p)
 	return err == nil
@@ -2698,19 +1112,6 @@ func (d *daemon) close() {
 	}
 	d.host.Close()
 	d.pdb.close()
-}
-
-// ---------------------------------------------------------------------------
-// XMP file write
-// ---------------------------------------------------------------------------
-
-func writeXMP(rawPath, content string) error {
-	xmpPath := rawPath + ".xmp"
-	tmp := xmpPath + ".tmp"
-	if err := os.WriteFile(tmp, []byte(content), 0644); err != nil {
-		return err
-	}
-	return os.Rename(tmp, xmpPath)
 }
 
 // pathMapPath returns the config-dir path for the persisted local→remote map.
